backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
tensorflow==2.18.0
pillow==10.2.0
python-jose==3.3.0
numpy==1.26.4
pydantic==2.6.3